refactor: isolate FileTrigger stories
